+++ src/components/pages/Roles/forms/RoleFormTabs.tsx
@@ -7,7 +7,7 @@ import { Hex, zeroAddress } from 'viem';
 import { DAO_ROUTES } from '../../../../constants/routes';
 import { useFractal } from '../../../../providers/App/AppProvider';
 import { useNetworkConfig } from '../../../../providers/NetworkConfig/NetworkConfigProvider';
-import { useRolesState } from '../../../../state/roles';
+import { useRolesStore } from '../../../../store/roles';
 import { EditBadgeStatus, EditedRole, RoleFormValues, RoleValue } from '../types';
 import RoleFormInfo from './RoleFormInfo';
 import RoleFormPaymentStream from './RoleFormPaymentStream';
@@ -20,7 +20,7 @@ const addRemoveField = (fieldNames: string[], fieldName: string, isRemoved: bool
 };
 
 export default function RoleFormTabs({ hatId, push }: { hatId: Hex; push: (obj: any) => void }) {
-  const { hatsTree } = useRolesState();
+  const { hatsTree } = useRolesStore();
   const {
     node: { daoAddress },
   } = useFractal();
