+++ src/pages/daos/[daoAddress]/roles/index.tsx
@@ -10,10 +10,10 @@ import PageHeader from '../../../../components/ui/page/Header/PageHeader';
 import { DAO_ROUTES } from '../../../../constants/routes';
 import { useFractal } from '../../../../providers/App/AppProvider';
 import { useNetworkConfig } from '../../../../providers/NetworkConfig/NetworkConfigProvider';
-import { useRolesState } from '../../../../state/roles';
+import { useRolesStore } from '../../../../store/roles';
 
 function Roles() {
-  const { hatsTree } = useRolesState();
+  const { hatsTree } = useRolesStore();
   const { addressPrefix } = useNetworkConfig();
   const { t } = useTranslation(['roles']);
   const {
